requirements.txt
PyQt5==5.15.7
matplotlib==3.5.2
reportlab==3.6.11
Pillow==9.2.0 